Add validation test utility.
OSG would like to be able to test if the appropriate Globus libaries
are available on the worker node; if not, we want to put the ones
we ship in a CVMFS repository into the environment of the auth
helper at configuration time.

This validation utility will return non-zero if the Globus/VOMS
install is broken.